[Bug #1634] Refactor the term loading even more to allow its usage in ParentTermSelector.
